added dominariworld

+++ programs/ecs/src/state.rs
@@ -2,9 +2,9 @@ use anchor_lang::prelude::*;
 
 #[derive(AnchorSerialize, AnchorDeserialize, Debug, Clone)]
 pub struct SerializedComponent{
-    pub component_key: Pubkey,
-    pub component_update_authority: Pubkey, // PDA of ComponentKey "update_authority"
+    pub component_key: Pubkey, // PDA from World Program
+    pub component_update_authority: Pubkey, // World Program usually
     pub schema_url: String, // Max 256 bytes ?
-    pub max_size: u64,
+    pub max_size: usize,
     pub data: Vec<u8>,
 }
